# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/student": {
      "post": {
        "operationId": "StudentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "StudentController_findAll",
        "parameters": [
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classSection",
            "required": false,
            "in": "query",
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "name": "firstName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lastName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pen",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateOfBirth",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/student/{id}": {
      "get": {
        "operationId": "StudentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "StudentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/school": {
      "post": {
        "operationId": "SchoolController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchoolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SchoolController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/School"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/school/completeAcademicYear": {
      "get": {
        "operationId": "SchoolController_completeAcademicYear",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/school/{id}": {
      "get": {
        "operationId": "SchoolController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/School"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "SchoolController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSchoolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/School"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SchoolController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "School"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/student-attendance": {
      "post": {
        "operationId": "StudentAttendanceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentAttendance"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Student Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "StudentAttendanceController_findAll",
        "parameters": [
          {
            "name": "studentId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "present",
                "absent",
                "late",
                "leave"
              ],
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "today",
                "week",
                "month"
              ],
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StudentAttendance"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Student Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/student-attendance/{id}": {
      "get": {
        "operationId": "StudentAttendanceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentAttendance"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Student Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "StudentAttendanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentAttendance"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Student Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentAttendanceController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Student Attendance"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "School Compass 365",
    "description": "swagger json: <a href=\"/swagger.json\">swagger.json</a>",
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "Mozilla Public License Version 2.0",
      "url": "https://github.com/vaibhavdubay/School-Compass-365-be/blob/main/LICENSE"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "establishedYear": {
            "type": "number"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicYear"
            }
          },
          "logo": {
            "$ref": "#/components/schemas/Image"
          },
          "logoUrl": {
            "type": "string"
          },
          "currentAcademicYear": {
            "$ref": "#/components/schemas/AcademicYear"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "schoolDISECode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Class"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "phoneNumber",
          "establishedYear",
          "address1",
          "address2",
          "academicYears",
          "logo",
          "logoUrl",
          "currentAcademicYear",
          "city",
          "state",
          "pincode",
          "schoolDISECode",
          "schoolCode",
          "classes",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other"
            ],
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "profileImage": {
            "$ref": "#/components/schemas/Image"
          },
          "profileImageUrl": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "school",
          "firstName",
          "lastName",
          "gender",
          "email",
          "phoneNumber",
          "profileImage",
          "profileImageUrl",
          "user",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "classId": {
            "type": "string"
          },
          "classSection": {
            "type": "string",
            "default": ""
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8
          },
          "userName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "pen": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "bloodGroup": {
            "type": "string",
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-"
            ]
          },
          "parentsGuardians": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParentOrGuardian"
            }
          }
        },
        "required": [
          "classId",
          "classSection",
          "firstName",
          "lastName",
          "email",
          "password",
          "userName",
          "phoneNumber",
          "pen",
          "dateOfBirth",
          "gender",
          "bloodGroup",
          "parentsGuardians"
        ]
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "class": {
            "$ref": "#/components/schemas/Class"
          },
          "classSection": {
            "type": "string",
            "default": ""
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "pen": {
            "type": "string"
          },
          "academicStatus": {
            "enum": [
              "active",
              "graduated",
              "transferred"
            ],
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other"
            ],
            "type": "string"
          },
          "bloodGroup": {
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-"
            ],
            "type": "string"
          },
          "profileImage": {
            "$ref": "#/components/schemas/Image"
          },
          "profileImageUrl": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "academicYears": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicYear"
            }
          },
          "parentsGuardians": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParentOrGuardian"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "school",
          "class",
          "classSection",
          "firstName",
          "lastName",
          "email",
          "phoneNumber",
          "pen",
          "academicStatus",
          "dateOfBirth",
          "gender",
          "bloodGroup",
          "profileImage",
          "profileImageUrl",
          "user",
          "academicYears",
          "parentsGuardians",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "classId": {
            "type": "string"
          },
          "classSection": {
            "type": "string",
            "default": ""
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8
          },
          "userName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "pen": {
            "type": "string"
          },
          "dateOfBirth": {
            "format": "date-time",
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          },
          "bloodGroup": {
            "type": "string",
            "enum": [
              "A+",
              "A-",
              "B+",
              "B-",
              "AB+",
              "AB-",
              "O+",
              "O-"
            ]
          },
          "parentsGuardians": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParentOrGuardian"
            }
          }
        }
      },
...
      "CreateSchoolDto": {
        "type": "object",
        "properties": {
          "schoolName": {
            "type": "string"
          },
          "establishedYear": {
            "type": "number"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "schoolDISECode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "classes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 8
          },
          "userName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ]
          }
        },
        "required": [
          "schoolName",
          "establishedYear",
          "address1",
          "city",
          "state",
          "pincode",
          "schoolDISECode",
          "schoolCode",
          "classes",
          "firstName",
          "lastName",
          "email",
          "password",
          "userName",
          "phoneNumber",
          "gender"
        ]
      },
      "UpdateSchoolDto": {
        "type": "object",
        "properties": {
          "image": {
            "type": "string",
            "format": "binary"
          },
          "name": {
            "type": "string"
          },
          "address1": {
            "type": "string"
          },
          "address2": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "pincode": {
            "type": "number"
          },
          "schoolDISECode": {
            "type": "string"
          },
          "schoolCode": {
            "type": "string"
          },
          "classes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address1",
          "city",
          "state",
          "pincode",
          "schoolDISECode",
          "schoolCode",
          "classes",
          "email",
          "phoneNumber"
        ]
      },
...
    }
  }
}